refactor: add eslint rules to linter exceptions
